requirements.txt
aiohttp
feedparser
Pillow
